pyproject.toml
[project]
name = "code-edditor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ansi2html>=1.9.2",
    "customtkinter>=5.2.2",
    "pip>=25.1.1",
    "pyside6>=6.9.0",
    "pywinpty>=2.0.15",
    "ttkbootstrap>=1.13.8",
]
